refactor: move projection types to core module
