+++ src/core/mod.rs
@@ -1,3 +1,10 @@
+import_all!(primitives);
+
 import_all!(error);
+
 import_all!(approximator);
 import_all!(parameterised);
+
+import_all!(projection);
+import_all!(projector);
+import_all!(feature);
